# swagger_server/models/response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Response(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, status: int=None, mesage: str=None, internal_transaction_id: str=None, external_transaction_id: str=None):  # noqa: E501
        """Response - a model defined in Swagger

        :param status: The status of this Response.  # noqa: E501
        :type status: int
        :param mesage: The mesage of this Response.  # noqa: E501
        :type mesage: str
        :param internal_transaction_id: The internal_transaction_id of this Response.  # noqa: E501
        :type internal_transaction_id: str
        :param external_transaction_id: The external_transaction_id of this Response.  # noqa: E501
        :type external_transaction_id: str
        """
        self.swagger_types = {
            'status': int,
            'mesage': str,
            'internal_transaction_id': str,
            'external_transaction_id': str
        }

        self.attribute_map = {
            'status': 'status',
            'mesage': 'mesage',
            'internal_transaction_id': 'internalTransactionId',
            'external_transaction_id': 'externalTransactionId'
        }
        self._status = status
        self._mesage = mesage
        self._internal_transaction_id = internal_transaction_id
        self._external_transaction_id = external_transaction_id

    @classmethod
    def from_dict(cls, dikt) -> 'Response':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Response of this Response.  # noqa: E501
        :rtype: Response
        """
        return util.deserialize_model(dikt, cls)

    @property
    def status(self) -> int:
        """Gets the status of this Response.


        :return: The status of this Response.
        :rtype: int
        """
        return self._status

    @status.setter
    def status(self, status: int):
        """Sets the status of this Response.


        :param status: The status of this Response.
        :type status: int
        """
        allowed_values = ["200", "400"]  # noqa: E501
        if status not in allowed_values:
            raise ValueError(
                "Invalid value for `status` ({0}), must be one of {1}"
                .format(status, allowed_values)
            )

        self._status = status

    @property
    def mesage(self) -> str:
        """Gets the mesage of this Response.


        :return: The mesage of this Response.
        :rtype: str
        """
        return self._mesage

    @mesage.setter
    def mesage(self, mesage: str):
        """Sets the mesage of this Response.


        :param mesage: The mesage of this Response.
        :type mesage: str
        """

        self._mesage = mesage

    @property
    def internal_transaction_id(self) -> str:
        """Gets the internal_transaction_id of this Response.


        :return: The internal_transaction_id of this Response.
        :rtype: str
        """
        return self._internal_transaction_id

    @internal_transaction_id.setter
    def internal_transaction_id(self, internal_transaction_id: str):
        """Sets the internal_transaction_id of this Response.


        :param internal_transaction_id: The internal_transaction_id of this Response.
        :type internal_transaction_id: str
        """

        self._internal_transaction_id = internal_transaction_id

    @property
    def external_transaction_id(self) -> str:
        """Gets the external_transaction_id of this Response.


        :return: The external_transaction_id of this Response.
        :rtype: str
        """
        return self._external_transaction_id

    @external_transaction_id.setter
    def external_transaction_id(self, external_transaction_id: str):
        """Sets the external_transaction_id of this Response.


        :param external_transaction_id: The external_transaction_id of this Response.
        :type external_transaction_id: str
        """

        self._external_transaction_id = external_transaction_id